+++ src/icemac/ab/calendar/calendar.py
@@ -56,19 +56,8 @@ class Calendar(zope.container.btree.BTreeContainer):
                            for x in recurring_events]
         events_map = {(x.category, x.in_timezone(timezone)): x
                       for x in itertools.chain(*recurred_events)}
-        catalog = zope.component.getUtility(zope.catalog.interfaces.ICatalog)
-        query = {
-            DATE_INDEX: {'between': (start, end, False, True)},
-            'schema_name': {'any_of': [
-                ISchemaName(IEvent).schema_name,
-                ISchemaName(IRecurringEvent).schema_name,
-            ]},
-        }
-        if categories:
-            query['keywords'] = {'any_of': categories}
-
-        # The values for the index are: min, max, min_exclude, max_exclude
-        single_events = catalog.searchResults(**query)
+        single_events = self.query_single_events(
+            start, end, categories=categories)
         # Sort deleted events first. This way a recurred event can be deleted
         # and later on replaced by a new event of the same category.
         sorted_single_events = sorted(
@@ -92,6 +81,21 @@ class Calendar(zope.container.btree.BTreeContainer):
         else:
             timezone = pytz.timezone(name)
         return timezone
+
+    def query_single_events(self, start, end, categories=[]):
+        catalog = zope.component.getUtility(zope.catalog.interfaces.ICatalog)
+        query = {
+            DATE_INDEX: {'between': (start, end, False, True)},
+            'schema_name': {'any_of': [
+                ISchemaName(IEvent).schema_name,
+                ISchemaName(IRecurringEvent).schema_name,
+            ]},
+        }
+        if categories:
+            query['keywords'] = {'any_of': categories}
+
+        # The values for the index are: min, max, min_exclude, max_exclude
+        return catalog.searchResults(**query)
 
 
 class CalendarDisplaySettings(grok.Annotation):